requirements.txt
Django==3.0.5
Pillow==7.1.1
PyYAML==5.3.1
dj-database-url==0.5.0
django-cors-headers==3.2.1
django-graphql-jwt==0.3.1
django-imagekit==4.0.2
django-parler==2.0.1
django-s3-storage==0.13.0
graphene-django==2.9.1
gunicorn==20.0.4
mypy==0.770
psycopg2==2.8.5
python-magic==0.4.15
